Life_AI/requirements.txt
fastapi
uvicorn[standard]
rdkit-pypi
python-multipart
numpy<2.0
